moving ObjectList to the components helper
